requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
pydantic
streamlit
google-generativeai
requests